src/main/java/guru/springframework/recipe/app/converters/fromdomain/UnitOfMeasureToUnitOfMeasureCommand.java
package guru.springframework.recipe.app.converters.fromdomain;

import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

import guru.springframework.recipe.app.commands.UnitOfMeasureCommand;
import guru.springframework.recipe.app.domain.UnitOfMeasure;
import lombok.Synchronized;

@Component
public class UnitOfMeasureToUnitOfMeasureCommand implements Converter<UnitOfMeasure, UnitOfMeasureCommand> {

	@Synchronized
	@Nullable
	@Override
	public UnitOfMeasureCommand convert(UnitOfMeasure source) {
		if (source == null) {
			return null;
		}
		
		UnitOfMeasureCommand destination = new UnitOfMeasureCommand();
		destination.setId(source.getId());
		destination.setDescription(source.getDescription());
		return destination;
	}

}
